requirements.txt
pandas
numpy
python-dateutil
sqlalchemy
alembic
phonenumbers
country_converter
validators
sqlparse

pre-commit
pytest
